+++ YYCalendar/Classes/Language.swift
@@ -1,0 +1,14 @@
+//
+//  Language.swift
+//  Pods
+//
+//  Created by DevYeom on 20/06/2019.
+//
+
+import Foundation
+
+let ENG_WEEK: [String] = ["SUN", "MON", "TUE", "WED", "THU", "FRI", "SAT"]
+let ENG2_WEEK: [String] = ["Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat"]
+let ENG3_WEEK: [String] = ["S", "M", "T", "W", "T", "F", "S"]
+let KOR_WEEK: [String] = ["일", "월", "화", "수", "목", "금", "토"]
+let CHN_WEEK: [String] = ["日", "月", "火", "水", "木", "金", "土"]
